Fix unit conversions.
The previous implementation was brittle and was likely to cause
problems for new module implementors.  The new implementation places
responsibility for parsing compound units on the individual modules,
instead of trying to do it generically.  Although it seems tempting to
try to factor this out into a generic implementation, there are just
too many idiosyncracies in the way that different variables compose
units.